requirements.txt
asgiref==3.7.2
Django==3.2.16
psycopg2==2.9.9
djangorestframework==3.14.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
PyJWT==2.8.0
PyYAML==6.0.1
uritemplate==4.1.1
